Refactor Test Utilities into Separate File
