+++ test_utils.py
@@ -1,0 +1,27 @@
+import pytest
+from unittest.mock import patch, MagicMock
+import os
+
+
+def mock_openai_chat_completion_create(*args, **kwargs):
+    class MockResponse:
+        def __init__(self):
+            self.choices = [self]
+
+        @property
+        def message(self):
+            return {"content": "Mocked Ad Headline"}
+
+    return MockResponse()
+
+
+@pytest.fixture(autouse=True)
+def set_openai_api_key():
+    with patch.dict(os.environ, {"OPENAI_API_KEY": "test"}):
+        yield
+
+
+def mock_openai_client():
+    mock_client = MagicMock()
+    mock_client.chat.completions.create.side_effect = mock_openai_chat_completion_create
+    return mock_client
